chore: move scratch files to examples/
