+++ examples/dogfood-simple.ts
@@ -1,4 +1,4 @@
-import { ClankaKernel } from './src/runtime/kernel.js';
+import { ClankaKernel } from '../src/runtime/kernel.js';
 import * as fs from 'node:fs';
 import * as path from 'node:path';
 import { createHash } from 'node:crypto';
