+++ examples/gen-golden.ts
